changed from int to bigint

+++ app/models.py
@@ -11,7 +11,7 @@ class Tweets(db.Model):
     hash_tag = db.Column(db.String(64),index=True)
     score=db.Column(db.Integer,default=0)
     date_created=db.Column(db.DateTime,index=True)
-    tweet_id=db.Column(db.Integer)
+    tweet_id=db.Column(db.BigInteger)
 
     def __repr__(self):
         return '<Tweet with hash_tag {} and score {}>'.format(self.hash_tag,self.score)
@@ -43,7 +43,7 @@ class Tweets(db.Model):
 
 class CursorPosition(db.Model):
     id=db.Column(db.Integer,primary_key=True)        
-    since_id=db.Column(db.Integer,default=1)
+    since_id=db.Column(db.BigInteger,default=1)
     key_word=db.Column(db.String(32))
 
     @staticmethod
